9 задание
