+++ js/ad.js
@@ -40,7 +40,7 @@ const generatorAd = function({offer,author}) {
     const tamplateAvatar = adElement.querySelector('img');
     tamplateAvatar.src = author.avatar;
   }
-  document.querySelector('#map-canvas').appendChild(adElement);
+  return adElement;
 };
